+++ Entregable-21/controllers/cartController.js
@@ -1,27 +1,27 @@
-const cartsDao = require('../daos/carts/index')
+const cartsApi = require('../api/cartsApi')
 
 const getCart = async (req, res) => {
     const { id } = req.params
-    res.json(await cartsDao.getCart(id))
+    res.json(await cartsApi.getCart(id))
 }
 
 const createCart = async (req, res) => {
-    res.json(await cartsDao.createCart())
+    res.json(await cartsApi.createCart())
 }
 
 const addItemToCart = async (req, res) => {
     const { id, id_prod} = req.params
-    res.json(await cartsDao.addItemToCart(id, id_prod))
+    res.json(await cartsApi.addItemToCart(id, id_prod))
 }
 
 const deleteCartItem = async (req, res) => {
     const { id, id_prod} = req.params
-    res.json(await cartsDao.deleteCartItem(id, id_prod))
+    res.json(await cartsApi.deleteCartItem(id, id_prod))
 }
 
 const deleteCart = async (req, res) => {
     const { id } = req.params
-    res.json(await cartsDao.deleteCart(id))
+    res.json(await cartsApi.deleteCart(id))
 }
 
 module.exports = {getCart, createCart, addItemToCart, deleteCartItem, deleteCart}
