--- Entregable-21/daos/purchases/index.js
@@ -1,1 +1,0 @@
-const a = 0;
