+++ Entregable-21/package.json
@@ -23,9 +23,11 @@
     "dotenv": "^16.0.3",
     "ejs": "^3.1.8",
     "express": "^4.18.1",
+    "express-graphql": "^0.12.0",
     "express-session": "^1.17.3",
     "faker": "^6.6.6",
     "firebase-admin": "^11.2.0",
+    "graphql": "^16.6.0",
     "knex": "^2.3.0",
     "log4js": "^6.7.1",
     "mongoose": "^6.7.1",
