+++ Entregable-21/routers/defaultRouter.js
@@ -1,5 +1,5 @@
 const yargs = require('yargs/yargs')(process.argv.slice(2))
-const { renderHome, renderProducts, renderCart, renderMessages, renderProfile, renderAdminPanel, renderFakeProducts, renderServerInfo, renderRandomNumbers, renderPurchases} = require('../controllers/defaultController')
+const { renderHome, renderCart, renderMessages, renderProfile, renderAdminPanel, renderFakeProducts, renderServerInfo, renderRandomNumbers, renderPurchases} = require('../controllers/defaultController')
 const { logRouteInfo } = require('../loggers/logger')
 const { Router } = require('express')
 const compression = require('compression')
@@ -9,9 +9,6 @@ const defaultRouter = Router()
 // ROUTER URL'S
 // Show the home screen
 defaultRouter.get('/', logRouteInfo, checkAuthentication, renderHome)
-
-// Show the products screen
-defaultRouter.get('/products', logRouteInfo, checkAuthentication, renderProducts)
 
 // Show the carts screen
 defaultRouter.get('/cart', logRouteInfo, checkAuthentication, renderCart)
