+++ Entregable-21/views/login.ejs
@@ -12,6 +12,7 @@
         <%- include ('partials/header.ejs') %>
         <div class = 'content-div'>
             <div class = 'column-centered-div expand'>
+                <h2> Login </h2>
                 <div class = 'general-card column-centered-div'>
                     <form action="/auth/login" method="post" class = "column-centered-div">
                         <input type="text" name = 'username' id = 'usernameInput' placeholder = "Username">
