+++ Entregable-21/views/partials/footer.ejs
@@ -1,6 +1,6 @@
 <footer>
     <div id = 'footer'>
         <h6 class = 'white'> The Summer Store ©</h6>
-        <p class = 'white'> Made with ❤️ in Argentina by <a href = 'twitter.com/santygegen' class = 'yellow'> Santy Gegenschatz </a> </p>
+        <p class = 'white'> Made with ❤️ in Argentina by <a href = 'https://santygegenschatz.netlify.app' class = 'yellow'> Santy Gegenschatz </a> </p>
     </div>
 </footer>
