+++ Entregable-21/views/signup.ejs
@@ -8,7 +8,7 @@
     <link rel="stylesheet" href="../../css/styles.css">
     <title> Signup </title>
 </head>
-<body>
+<body id = 'wrapper'>
     <%- include ('partials/header.ejs')%>
     <div class = 'content-div'>
         <div class = 'general-card column-centered-div'>
